+++ tests/test-data.R
@@ -371,6 +371,7 @@ test_that("Filter bots from commit data", {
     filtered.commits = proj.data$get.commits()
 
     expect_true(all(filtered.commits[["author.name"]] != "Thomas"))
+    ## there are now 5 commits remaining since 3 commits got filtered
     expect_equal(nrow(filtered.commits), 5)
 })
 
@@ -385,6 +386,7 @@ test_that("Filter bots from issue data", {
     filtered.issues = proj.data$get.issues()
 
     expect_true(all(filtered.issues[["author.name"]] != "Thomas"))
+    ## there are now 41 issues remaining since 6 issues got filtered
     expect_equal(nrow(filtered.issues), 41)
 })
 
@@ -399,6 +401,7 @@ test_that("Filter bots from mail data", {
     filtered.mails = proj.data$get.mails()
 
     expect_true(all(filtered.mails[["author.name"]] != "Thomas"))
+    ## there are now 15 mails remaining since 2 mails got filtered
     expect_equal(nrow(filtered.mails), 15)
 })