+++ util-read.R
@@ -389,15 +389,13 @@ create.empty.issues.list = function() {
 
 
 ## column names of a data frame containing bot information (see file
-## 'TODO-bots.list' and function \code{read.bot.list})
-## TODO figure out what the actual columns will be
+## 'bots.list' and function \code{read.bot.list})
 BOT.LIST.COLUMNS = c(
     "author.name", "author.email", ## author
-    "comment.count", "comment.empty.count", "comment.pattern.count", "comment.dispersion", ## intermediate results
     "is.bot" ## whether this is a bot
 )
 
-#' Read the bot classification from the 'TODO-bots.list' file.
+#' Read the bot classification from the 'bots.list' file.
 #'
 #' @param data.path the path to the commit-messages list
 #'
@@ -406,15 +404,15 @@ BOT.LIST.COLUMNS = c(
 read.bot.info = function(data.path) {
     logging::logdebug("read.bot.info: starting.")
 
-    ## read the file with the commit messages
-    file = file.path(data.path, "TODO-bots.list")
+    ## read the file with the bot info
+    file = file.path(data.path, "bots.list")
 
     bot.data = try(read.table(file, header = FALSE, sep = ";", strip.white = TRUE,
                                          encoding = "UTF-8"), silent = TRUE)
 
-    ## handle the case that the list of commits is empty
+    ## handle the case that the bot info is empty
     if (inherits(bot.data, "try-error")) {
-        logging::logwarn("There is no bot clarification available for the current environment.")
+        logging::logwarn("There is no bot information available for the current environment.")
         logging::logwarn("Datapath: %s", data.path)
 
         ## return a data frame with the correct columns but zero rows
@@ -423,13 +421,8 @@ read.bot.info = function(data.path) {
 
     ## set column names for new data frame
     colnames(bot.data) = BOT.LIST.COLUMNS
-
-    bot.data = bot.data[, c("author.name", "author.email", "is.bot")]
     bot.data["is.bot"] = sapply(bot.data[["is.bot"]], function(x) switch(x, Bot = TRUE, Human = FALSE, NA))
-
-
     logging::logdebug("read.bot.info: finished.")
-
     return(bot.data)
 }
 
@@ -475,7 +468,6 @@ read.authors = function(data.path) {
 
     bot.data = read.bot.info(data.path)
     if (!is.null(bot.data)) {
-        ## TODO keep row names consistent
         authors.df = merge(authors.df, bot.data, by = c("author.name", "author.email"), all.x = TRUE, sort = FALSE)
         authors.df = authors.df[order(authors.df[["author.id"]]),] # re-order after read
         row.names(authors.df) = 1:nrow(authors.df)